server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.1.4
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.0
numpy==1.24.3
aiofiles==23.2.1